requirements.txt
azure-functions
tensorflow==2.0.0
Pillow
requests
